fixed that it wasn't correctly showing chance of rain

+++ app/src/main/java/net/rmoreno/weatherapp/CurrentWeather.java
@@ -13,11 +13,11 @@ public class CurrentWeather {
     String mIcon;
     String mSummary;
     String mTimeZone;
-    int mPrecipitation;
+    double mPrecipitation;
     double mWindSpeed;
     double mFeelsLike;
 
-    public void setPrecip(int precipitation){
+    public void setPrecip(double precipitation){
         mPrecipitation = precipitation;
     }
